5 test cases and solution
