--- RUAP_LV4.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30804.86
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RUAP_LV4", "RUAP_LV4\RUAP_LV4.csproj", "{C37FDBF6-E402-4429-9216-2B976FF3663E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C37FDBF6-E402-4429-9216-2B976FF3663E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C37FDBF6-E402-4429-9216-2B976FF3663E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C37FDBF6-E402-4429-9216-2B976FF3663E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C37FDBF6-E402-4429-9216-2B976FF3663E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6B613CA9-350B-4589-A63A-120EDB5C689C}
-	EndGlobalSection
-EndGlobal
