+++ RUAP_LV4/packages.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MSTest.TestAdapter" version="2.1.1" targetFramework="net472" />
+  <package id="MSTest.TestFramework" version="2.1.1" targetFramework="net472" />
+  <package id="NUnit" version="3.12.0" targetFramework="net472" />
+  <package id="NUnit3TestAdapter" version="3.15.1" targetFramework="net472" />
+  <package id="Selenium.Support" version="3.141.0" targetFramework="net472" />
+  <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net472" />
+  <package id="Selenium.WebDriver.GeckoDriver" version="0.26.0" targetFramework="net472" />
+</packages>
