+++ katalon/TestCase_ProvjeraDostave.htm
@@ -1,0 +1,116 @@
+using System;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading;
+using NUnit.Framework;
+using OpenQA.Selenium;
+using OpenQA.Selenium.Firefox;
+using OpenQA.Selenium.Support.UI;
+
+namespace SeleniumTests
+{
+    [TestFixture]
+    public class UntitledTestCase
+    {
+        private IWebDriver driver;
+        private StringBuilder verificationErrors;
+        private string baseURL;
+        private bool acceptNextAlert = true;
+        
+        [SetUp]
+        public void SetupTest()
+        {
+            driver = new FirefoxDriver();
+            baseURL = "https://www.google.com/";
+            verificationErrors = new StringBuilder();
+        }
+        
+        [TearDown]
+        public void TeardownTest()
+        {
+            try
+            {
+                driver.Quit();
+            }
+            catch (Exception)
+            {
+                // Ignore errors if unable to close the browser
+            }
+            Assert.AreEqual("", verificationErrors.ToString());
+        }
+        
+        [Test]
+        public void TheUntitledTestCaseTest()
+        {
+            driver.Navigate().GoToUrl("https://demo.opencart.com/index.php?route=common/home");
+            driver.FindElement(By.XPath("//header")).Click();
+            driver.FindElement(By.XPath("//div[@id='top-links']/ul/li")).Click();
+            driver.FindElement(By.XPath("//div[@id='top-links']/ul/li[2]/a/i")).Click();
+            driver.FindElement(By.LinkText("Login")).Click();
+            driver.FindElement(By.Id("input-email")).Click();
+            driver.FindElement(By.Id("input-email")).Clear();
+            driver.FindElement(By.Id("input-email")).SendKeys("fakemail@gmail.com");
+            driver.FindElement(By.Id("input-password")).Click();
+            driver.FindElement(By.Id("input-password")).Clear();
+            driver.FindElement(By.Id("input-password")).SendKeys("jasamzvone123");
+            driver.FindElement(By.Id("input-password")).SendKeys(Keys.Enter);
+            driver.FindElement(By.LinkText("Tablets")).Click();
+            driver.FindElement(By.XPath("(//button[@type='button'])[9]")).Click();
+            driver.FindElement(By.Id("cart-total")).Click();
+            driver.FindElement(By.XPath("//div[@id='cart']/ul/li[2]/div/p/a/strong")).Click();
+            driver.FindElement(By.LinkText("Estimate Shipping & Taxes")).Click();
+            driver.FindElement(By.Id("input-country")).Click();
+            new SelectElement(driver.FindElement(By.Id("input-country"))).SelectByText("Croatia");
+            driver.FindElement(By.Id("input-zone")).Click();
+            new SelectElement(driver.FindElement(By.Id("input-zone"))).SelectByText("Osječko-baranjska");
+            driver.FindElement(By.Id("input-zone")).Click();
+            driver.FindElement(By.Id("input-postcode")).Click();
+            driver.FindElement(By.Id("input-postcode")).Clear();
+            driver.FindElement(By.Id("input-postcode")).SendKeys("31000");
+            driver.FindElement(By.XPath("//div[@id='collapse-shipping']/div/div")).Click();
+            driver.FindElement(By.Id("button-quote")).Click();
+            driver.FindElement(By.XPath("//div[@id='modal-shipping']/div/div/div[2]/div/label")).Click();
+            driver.FindElement(By.Id("button-shipping")).Click();
+        }
+        private bool IsElementPresent(By by)
+        {
+            try
+            {
+                driver.FindElement(by);
+                return true;
+            }
+            catch (NoSuchElementException)
+            {
+                return false;
+            }
+        }
+        
+        private bool IsAlertPresent()
+        {
+            try
+            {
+                driver.SwitchTo().Alert();
+                return true;
+            }
+            catch (NoAlertPresentException)
+            {
+                return false;
+            }
+        }
+        
+        private string CloseAlertAndGetItsText() {
+            try {
+                IAlert alert = driver.SwitchTo().Alert();
+                string alertText = alert.Text;
+                if (acceptNextAlert) {
+                    alert.Accept();
+                } else {
+                    alert.Dismiss();
+                }
+                return alertText;
+            } finally {
+                acceptNextAlert = true;
+            }
+        }
+    }
+}
